organize test declarations
