+++ test/tests/interfaces/main.c
@@ -1,6 +1,8 @@
 #include "sicgl/screen.h"
 #include "utils.h"
 
+#include "tests.h"
+
 void setUp(void) {
   // set stuff up here
 }
@@ -9,256 +11,15 @@ void tearDown(void) {
   // clean stuff up here
 }
 
-void test_specific_display_pixel(void) {
-  uext_t width = 3;
-  uext_t height = 3;
-
-  gdImage* reference;
-  gdImage* image;
-  display_t* display;
-  specific_interface_t* interface;
-  color_sequence_t* color_sequence;
-  png_t* ref;
-  png_t* img;
-
-  if (TEST_PROTECT()) {
-    // create images
-    reference = new_image(width, height);
-    TEST_ASSERT_NOT_NULL_MESSAGE(
-        reference, "could not allocate reference image");
-
-    // create specific interface
-    display = new_display(width, height, 0, 0);
-    interface = new_libgd_specific_interface(display, NULL, 0);
-    TEST_ASSERT_NOT_NULL_MESSAGE(display, "could not create display");
-    TEST_ASSERT_NOT_NULL_MESSAGE(interface, "could not create interface");
-
-    // create a color sequence
-    int color = truecolor_random_color();
-    color_sequence = new_color_sequence(
-        SICGL_CS_STATIC, sizeof(int), (uint8_t*)&color, sizeof(color));
-    TEST_ASSERT_NOT_NULL_MESSAGE(
-        color_sequence, "could not create color sequence");
-
-    // draw a pixel
-    int ret = simultaneous_specific_display_pixel(
-        reference, interface, color_sequence, 1, 1);
-    image = new_image_from_libgd_specific_interface(interface);
-    TEST_ASSERT_NOT_NULL_MESSAGE(image, "could not convert test image");
-
-    // save images to png
-    ref = new_png_from_image(reference);
-    img = new_png_from_libgd_specific_interface(interface);
-    TEST_ASSERT_NOT_NULL_MESSAGE(ref, "could not create ref png");
-    TEST_ASSERT_NOT_NULL_MESSAGE(img, "could not create img png");
-    TEST_ASSERT_EQUAL_INT(
-        0, png_to_file(ref, TEST_OUTPUT_DIR "/specific_display_pixel_ref.png"));
-    TEST_ASSERT_EQUAL_INT(
-        0, png_to_file(img, TEST_OUTPUT_DIR "/specific_display_pixel_img.png"));
-
-    // compare the images
-    TEST_ASSERT_EQUAL_INT(0, compare_image(reference, image));
-
-  } else {
-    release_image(reference);
-    release_image(image);
-    release_display(display);
-    release_libgd_specific_interface(interface);
-    release_color_sequence(color_sequence);
-    release_png(ref);
-    release_png(img);
-  }
-}
-
-void test_specific_display_hline(void) {
-  uext_t width = 30;
-  uext_t height = 3;
-
-  gdImage* reference;
-  gdImage* image;
-  display_t* display;
-  specific_interface_t* interface;
-  color_sequence_t* color_sequence;
-  png_t* ref;
-  png_t* img;
-
-  if (TEST_PROTECT()) {
-    // create images
-    reference = new_image(width, height);
-    TEST_ASSERT_NOT_NULL_MESSAGE(
-        reference, "could not allocate reference image");
-
-    // create specific interface
-    display = new_display(width, height, 0, 0);
-    interface = new_libgd_specific_interface(display, NULL, 0);
-    TEST_ASSERT_NOT_NULL_MESSAGE(display, "could not create display");
-    TEST_ASSERT_NOT_NULL_MESSAGE(interface, "could not create interface");
-
-    // create a color sequence
-    int color = truecolor_random_color();
-    color_sequence = new_color_sequence(
-        SICGL_CS_STATIC, sizeof(int), (uint8_t*)&color, sizeof(color));
-    TEST_ASSERT_NOT_NULL_MESSAGE(
-        color_sequence, "could not create color sequence");
-
-    // draw a pixel
-    int ret = simultaneous_specific_display_hline(
-        reference, interface, color_sequence, 1, 1, 28, 1);
-    image = new_image_from_libgd_specific_interface(interface);
-    TEST_ASSERT_NOT_NULL_MESSAGE(image, "could not convert test image");
-
-    // save images to png
-    ref = new_png_from_image(reference);
-    img = new_png_from_libgd_specific_interface(interface);
-    TEST_ASSERT_NOT_NULL_MESSAGE(ref, "could not create ref png");
-    TEST_ASSERT_NOT_NULL_MESSAGE(img, "could not create img png");
-    TEST_ASSERT_EQUAL_INT(
-        0, png_to_file(ref, TEST_OUTPUT_DIR "/specific_display_hline_ref.png"));
-    TEST_ASSERT_EQUAL_INT(
-        0, png_to_file(img, TEST_OUTPUT_DIR "/specific_display_hline_img.png"));
-
-    // compare the images
-    TEST_ASSERT_EQUAL_INT(0, compare_image(reference, image));
-
-  } else {
-    release_image(reference);
-    release_image(image);
-    release_display(display);
-    release_libgd_specific_interface(interface);
-    release_color_sequence(color_sequence);
-    release_png(ref);
-    release_png(img);
-  }
-}
-
-void test_specific_display_vline(void) {
-  uext_t width = 3;
-  uext_t height = 30;
-
-  gdImage* reference;
-  gdImage* image;
-  display_t* display;
-  specific_interface_t* interface;
-  color_sequence_t* color_sequence;
-  png_t* ref;
-  png_t* img;
-
-  if (TEST_PROTECT()) {
-    // create images
-    reference = new_image(width, height);
-    TEST_ASSERT_NOT_NULL_MESSAGE(
-        reference, "could not allocate reference image");
-
-    // create specific interface
-    display = new_display(width, height, 0, 0);
-    interface = new_libgd_specific_interface(display, NULL, 0);
-    TEST_ASSERT_NOT_NULL_MESSAGE(display, "could not create display");
-    TEST_ASSERT_NOT_NULL_MESSAGE(interface, "could not create interface");
-
-    // create a color sequence
-    int color = truecolor_random_color();
-    color_sequence = new_color_sequence(
-        SICGL_CS_STATIC, sizeof(int), (uint8_t*)&color, sizeof(color));
-    TEST_ASSERT_NOT_NULL_MESSAGE(
-        color_sequence, "could not create color sequence");
-
-    // draw a pixel
-    int ret = simultaneous_specific_display_vline(
-        reference, interface, color_sequence, 1, 1, 1, 28);
-    image = new_image_from_libgd_specific_interface(interface);
-    TEST_ASSERT_NOT_NULL_MESSAGE(image, "could not convert test image");
-
-    // save images to png
-    ref = new_png_from_image(reference);
-    img = new_png_from_libgd_specific_interface(interface);
-    TEST_ASSERT_NOT_NULL_MESSAGE(ref, "could not create ref png");
-    TEST_ASSERT_NOT_NULL_MESSAGE(img, "could not create img png");
-    TEST_ASSERT_EQUAL_INT(
-        0, png_to_file(ref, TEST_OUTPUT_DIR "/specific_display_vline_ref.png"));
-    TEST_ASSERT_EQUAL_INT(
-        0, png_to_file(img, TEST_OUTPUT_DIR "/specific_display_vline_img.png"));
-
-    // compare the images
-    TEST_ASSERT_EQUAL_INT(0, compare_image(reference, image));
-
-  } else {
-    release_image(reference);
-    release_image(image);
-    release_display(display);
-    release_libgd_specific_interface(interface);
-    release_color_sequence(color_sequence);
-    release_png(ref);
-    release_png(img);
-  }
-}
-
-void test_specific_display_region(void) {
-  uext_t width = 30;
-  uext_t height = 30;
-
-  gdImage* reference;
-  gdImage* image;
-  display_t* display;
-  specific_interface_t* interface;
-  color_sequence_t* color_sequence;
-  png_t* ref;
-  png_t* img;
-
-  if (TEST_PROTECT()) {
-    // create images
-    reference = new_image(width, height);
-    TEST_ASSERT_NOT_NULL_MESSAGE(
-        reference, "could not allocate reference image");
-
-    // create specific interface
-    display = new_display(width, height, 0, 0);
-    interface = new_libgd_specific_interface(display, NULL, 0);
-    TEST_ASSERT_NOT_NULL_MESSAGE(display, "could not create display");
-    TEST_ASSERT_NOT_NULL_MESSAGE(interface, "could not create interface");
-
-    // create a color sequence
-    int color = truecolor_random_color();
-    color_sequence = new_color_sequence(
-        SICGL_CS_STATIC, sizeof(int), (uint8_t*)&color, sizeof(color));
-    TEST_ASSERT_NOT_NULL_MESSAGE(
-        color_sequence, "could not create color sequence");
-
-    // draw a pixel
-    int ret = simultaneous_specific_display_region(
-        reference, interface, color_sequence, 1, 1, 28, 28);
-    image = new_image_from_libgd_specific_interface(interface);
-    TEST_ASSERT_NOT_NULL_MESSAGE(image, "could not convert test image");
-
-    // save images to png
-    ref = new_png_from_image(reference);
-    img = new_png_from_libgd_specific_interface(interface);
-    TEST_ASSERT_NOT_NULL_MESSAGE(ref, "could not create ref png");
-    TEST_ASSERT_NOT_NULL_MESSAGE(img, "could not create img png");
-    TEST_ASSERT_EQUAL_INT(
-        0, png_to_file(ref, TEST_OUTPUT_DIR "/specific_display_region_ref.png"));
-    TEST_ASSERT_EQUAL_INT(
-        0, png_to_file(img, TEST_OUTPUT_DIR "/specific_display_region_img.png"));
-
-    // compare the images
-    TEST_ASSERT_EQUAL_INT(0, compare_image(reference, image));
-
-  } else {
-    release_image(reference);
-    release_image(image);
-    release_display(display);
-    release_libgd_specific_interface(interface);
-    release_color_sequence(color_sequence);
-    release_png(ref);
-    release_png(img);
-  }
-}
-
 // not needed when using generate_test_runner.rb
 int main(void) {
   UNITY_BEGIN();
+
+	// specific display tests
   RUN_TEST(test_specific_display_pixel);
   RUN_TEST(test_specific_display_hline);
   RUN_TEST(test_specific_display_vline);
 	RUN_TEST(test_specific_display_region);
+
   return UNITY_END();
 }
