+++ test/tests/interfaces/tests.h
@@ -1,0 +1,13 @@
+#pragma once
+
+// specific display
+void test_specific_display_pixel(void);
+void test_specific_display_hline(void);
+void test_specific_display_vline(void);
+void test_specific_display_region(void);
+
+// specific screen
+void test_specific_screen_pixel(void);
+void test_specific_screen_hline(void);
+void test_specific_screen_vline(void);
+void test_specific_screen_region(void);
